adds Altair notification for BC node operators to eth2 pages
