+++ src/pages/eth2/index.js
@@ -15,7 +15,7 @@ import GhostCard from "../../components/GhostCard"
 import InfoBanner from "../../components/InfoBanner"
 import Link from "../../components/Link"
 import PageMetadata from "../../components/PageMetadata"
-import BannerNotification from "../../components/BannerNotification"
+import UpgradeBannerNotification from "../../components/UpgradeBannerNotification"
 import Translation from "../../components/Translation"
 import PageHero from "../../components/PageHero"
 import {
@@ -202,16 +202,6 @@ const ResearchContainer = styled.div`
   margin-top: 2rem;
 `
 
-const StyledBannerNotification = styled(BannerNotification)`
-  display: flex;
-  justify-content: center;
-`
-
-const StyledEmoji = styled(Emoji)`
-  margin-right: 1rem;
-  flex-shrink: 0;
-`
-
 const paths = [
   {
     emoji: ":rocket:",
@@ -282,6 +272,7 @@ const Eth2IndexPage = ({ data }) => {
         title={translateMessageId("page-eth2-meta-title", intl)}
         description={translateMessageId("page-eth2-meta-desc", intl)}
       />
+      <UpgradeBannerNotification />
       <PageHero content={heroContent} />
       <Divider />
       <Content>
