+++ src/pages/eth2/vision.js
@@ -14,7 +14,7 @@ import PageHero from "../../components/PageHero"
 import Breadcrumbs from "../../components/Breadcrumbs"
 import ButtonLink from "../../components/ButtonLink"
 import PageMetadata from "../../components/PageMetadata"
-import BannerNotification from "../../components/BannerNotification"
+import UpgradeBannerNotification from "../../components/UpgradeBannerNotification"
 import {
   CardContainer,
   Content,
@@ -48,16 +48,6 @@ const CentreCard = styled(Card)`
   }
 `
 
-const StyledCard = styled(Card)`
-  flex: 1 1 30%;
-  min-width: 240px;
-  margin: 1rem;
-  padding: 1.5rem;
-  @media (max-width: ${(props) => props.theme.breakpoints.l}) {
-    flex: 1 1 30%;
-  }
-`
-
 const CentralContent = styled.div`
   margin: 0rem 12rem;
   justify-content: center;
@@ -75,16 +65,6 @@ const TrilemmaContent = styled.div`
 
 const StyledBreadcrumbs = styled(Breadcrumbs)`
   justify-content: center;
-`
-
-const StyledBannerNotification = styled(BannerNotification)`
-  display: flex;
-  justify-content: center;
-`
-
-const StyledEmoji = styled(Emoji)`
-  margin-right: 1rem;
-  flex-shrink: 0;
 `
 
 const paths = [
@@ -146,6 +126,7 @@ const VisionPage = ({ data, location }) => {
         title={translateMessageId("page-eth2-vision-meta-title", intl)}
         description={translateMessageId("page-eth2-vision-meta-desc", intl)}
       />
+      <UpgradeBannerNotification />
       <PageHero content={heroContent} />
       <Divider />
       <Content>
